Create README.md
The challenge of this project is to predict if the car purchased at the Auction is bad buy. One of the biggest challenges of an auto dealership purchasing a used car at an auto auction is the risk of that the vehicle might have serious issues that prevent it from being sold to customers. The auto community calls these unfortunate purchases "kicks". Kicked cars often result when there are tampered odometers, mechanical issues the dealer is not able to address, issues with getting the vehicle title from the seller, or some other unforeseen problem. Kick cars can be very costly to dealers after transportation cost, throwaway repair work, and market losses in reselling the vehicle. Modelers who can figure out which cars have a higher risk of being kicked can provide real value to dealerships trying to provide the best inventory selection possible to their customers.